separate cluster and node
